Toxicity Rating Model/app/requirements.txt
Flask==2.2.5
pandas==2.1.4
numpy==1.26.2
# tensorflow==2.10.0
tensorflow-gpu==2.10.0
